Refactor privacy modal logic to work with latest version of login form
